% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{linearRegressionApp}
\alias{linearRegressionApp}
\title{Interactive Linear Regression Shiny App}
\usage{
linearRegressionApp()
}
\value{
This function launches a Shiny app and does not return any value.
}
\description{
This function launches a Shiny app for performing interactive linear regression analysis.
The app allows users to upload a CSV file, choose various options for data preprocessing,
perform exploratory data analysis, fit linear regression models, conduct model diagnostics,
perform stepwise model selection, and compare different models.
}
\examples{
\dontrun{
linearRegressionApp()
}
}
\author{
Adam Janczyszyn, Jakub Zmujdzin
}
